Added timestamps to databse and fixed some endpoints

+++ routes/post.js
@@ -134,19 +134,55 @@ router.put("/unlike/:postId", restrictUser, async (req, res) => {
 // PUT /post/:id
 router.put("/:id", restrictUser, async (req, res) => {
   try {
+    const loggedInUserId = req.session.user.id;
     const postId = req.params.id;
-    const { description, resources, userId } = req.body;
+    const { description, userId } = req.body;
 
-    const post = await prisma.post.update({
+    const postUser = await prisma.post.findUnique({
       where: {
         id: Number(postId),
       },
-      data: {
-        description: description,
-        resources: resources,
-        userId: Number(userId),
+      select: {
+        userId: true,
       },
     });
+
+    if (!description) {
+      return res.status(400).json({ message: "Description is required" });
+    }
+    if (
+      postUser.userId !== loggedInUserId &&
+      req.session.user.role.role != "admin"
+    ) {
+      return res
+        .status(403)
+        .json({ message: "You are not authorized to edit this post" });
+    }
+
+    let post;
+
+    if (!userId) {
+      post = await prisma.post.update({
+        where: {
+          id: Number(postId),
+        },
+        data: {
+          description: description,
+          updatedAt: new Date(),
+        },
+      });
+    } else {
+      post = await prisma.post.update({
+        where: {
+          id: Number(postId),
+        },
+        data: {
+          description: description,
+          userId: userId,
+          updatedAt: new Date(),
+        },
+      });
+    }
     res.status(200).json(post);
   } catch (err) {
     console.log(err);
@@ -157,14 +193,33 @@ router.put("/:id", restrictUser, async (req, res) => {
 //DELETE /post/:id
 router.delete("/:id", restrictUser, async (req, res) => {
   try {
-    const id = req.params.id;
-    console.log(id);
-    const post = await prisma.post.delete({
+    const loggedInUserId = req.session.user.id;
+    const postId = req.params.id;
+
+    const postUser = await prisma.post.findUnique({
       where: {
-        id: Number(id),
+        id: Number(postId),
+      },
+      select: {
+        userId: true,
       },
     });
-    res.sendStatus(204);
+
+    if (
+      postUser.userId !== loggedInUserId &&
+      req.session.user.role.role != "admin"
+    ) {
+      return res
+        .status(403)
+        .json({ message: "You are not authorized to delete this post" });
+    }
+
+    const post = await prisma.post.delete({
+      where: {
+        id: Number(postId),
+      },
+    });
+    res.status(200).json(post);
   } catch (err) {
     console.log(err);
     return res.sendStatus(500);
